Fix until so it works >0x80000000

+++ src/devices/dev_mp.cc
@@ -64,7 +64,7 @@ struct mp_data {
 };
 
 
-extern int single_step;
+extern uint64_t single_step;
 
 
 DEVICE_ACCESS(mp)
